backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
email-validator
requests

# HTTP client
httpx==0.26.0

# Environment variables
python-dotenv==1.0.0

# OpenAI
openai==1.10.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Linting and Formatting
flake8==7.0.0
black==24.1.0
